Created logs system

+++ system/Controllers/TicketsController.php
@@ -12,18 +12,20 @@ class TicketsController extends Controller
 {
     use ValidateTicket;
     private $ticketModel;
-    private $privileges;
     private $categoryModel;
     private $userModel;
+    private $logModel;
+    private $privileges;
 
     public function __construct()
     {
         global $lang;
 
-        // load the model
+        // load models
         $this->ticketModel = $this->loadModel('Ticket');
         $this->categoryModel = $this->loadModel('Category');
         $this->userModel = $this->loadModel('User');
+        $this->logModel = $this->loadModel('Log');
 
         // store user privileges
         $this->privileges = $this->checkPrivileges();
@@ -156,8 +158,16 @@ class TicketsController extends Controller
             $category_name = $this->ticketModel->getCategoryName($ticket['category_id']);
 
             // check if delete group button is set
-            if (isset($_POST['delete_ticket'])) {
+            if (in_array(1, $this->privileges['canDeleteTickets']) && isset($_POST['delete_ticket'])) {
                 if ($this->ticketModel->deleteTicket($id)) {
+                    // log action
+                    $logAction = $_SESSION['user_name'] . ' deleted ticket (ID: ' . $id . ').';
+                    $logData = [
+                        'type' => 'Ticket',
+                        'action' => $logAction
+                    ];
+                    $this->logModel->adminLog($logData);
+
                     flashMessage('success', 'Ticket has been successfully deleted!');
                     redirect('/tickets');
                     unset($_POST);
@@ -167,11 +177,19 @@ class TicketsController extends Controller
             }
 
             // check if close group button is set
-            if (isset($_POST['close_ticket'])) {
+            if (in_array(1, $this->privileges['canCloseTickets']) && isset($_POST['close_ticket'])) {
                 $status = 'Closed';
                 $closed_by = $_SESSION['user_id'];
 
                 if ($this->ticketModel->updateStatus($id, $status, $closed_by)) {
+                    // log action
+                    $logAction = $_SESSION['user_name'] . ' closed ticket (ID: ' . $id . ').';
+                    $logData = [
+                        'type' => 'Ticket',
+                        'action' => $logAction
+                    ];
+                    $this->logModel->adminLog($logData);
+
                     flashMessage('success', 'Ticket has been successfully closed!');
                     redirect('/tickets');
                     unset($_POST);
@@ -193,52 +211,66 @@ class TicketsController extends Controller
                 'badges' => $badges
             ];
 
-            if (isset($_POST['edit_ticket'])) {
-                // sanitize post data
-                $_POST['ticket_category'] = filter_var($_POST['ticket_category'], FILTER_SANITIZE_NUMBER_INT);
-                $_POST['ticket_body'] = htmlentities($_POST['ticket_body']);
+            if ($_SESSION['user_id'] == $ticket['author_id'] || in_array(1, $this->privileges['canEditTickets'])) {
+                if (isset($_POST['edit_ticket'])) {
+                    // sanitize post data
+                    $_POST['ticket_category'] = filter_var($_POST['ticket_category'], FILTER_SANITIZE_NUMBER_INT);
+                    $_POST['ticket_body'] = htmlentities($_POST['ticket_body']);
 
-                $dataPost = [
-                    'body' => $_POST['ticket_body'],
-                    'category_id' => $_POST['ticket_category'],
-                    'edit_ip' => getUserIp()
-                ];
+                    $dataPost = [
+                        'body' => $_POST['ticket_body'],
+                        'category_id' => $_POST['ticket_category'],
+                        'edit_ip' => getUserIp()
+                    ];
 
-                // handle errors
-                $errors = ValidateTicket::validate($dataPost);
+                    // handle errors
+                    $errors = ValidateTicket::validate($dataPost);
 
-                // check if there are no errors
-                if (count(array_filter($errors)) == 0) {
-                    // create group
-                    if ($this->ticketModel->editTicket($dataPost, $id)) {
-                        flashMessage('success', 'The ticket has been successfully edited!');
-                        redirect('/tickets');
+                    // check if there are no errors
+                    if (count(array_filter($errors)) == 0) {
+                        // create group
+                        if ($this->ticketModel->editTicket($dataPost, $id)) {
+                            // log action
+                            $logAction = $_SESSION['user_name'] . ' edited ticket (ID: ' . $id . ').';
+                            $logData = [
+                                'type' => 'Ticket',
+                                'action' => $logAction
+                            ];
+                            if ($_SESSION['user_id'] == $ticket['author_id']) {
+                                $this->logModel->playerLog($logData);
+                            } else {
+                                $this->logModel->adminLog($logData);
+                            }
+
+                            flashMessage('success', 'The ticket has been successfully edited!');
+                            redirect('/tickets');
+                        } else {
+                            die('Something went wrong.');
+                        }
                     } else {
-                        die('Something went wrong.');
+                        // load view with errors
+                        $this->loadView('ticket_edit', $data, $errors);
                     }
                 } else {
-                    // load view with errors
-                    $this->loadView('ticket_edit', $data, $errors);
-                }
-            } else {
-                $data = [
-                    'pageTitle' => 'Edit Ticket',
-                    'fullAccess' => $this->privileges['fullAccess'],
-                    'isAdmin' => $this->privileges['isAdmin'],
-                    'isLeader' => $this->privileges['isLeader'],
-                    'canDeleteTickets' => in_array(1, $this->privileges['canDeleteTickets']),
-                    'ticket' => [
-                        'body' => $ticket['body'],
-                        'category_id' => $ticket['category_id']
-                    ],
-                    'category_name' => $category_name,
-                    'categories' => $categories,
-                    'lang' => $lang,
-                    'badges' => $badges
-                ];
+                    $data = [
+                        'pageTitle' => 'Edit Ticket',
+                        'fullAccess' => $this->privileges['fullAccess'],
+                        'isAdmin' => $this->privileges['isAdmin'],
+                        'isLeader' => $this->privileges['isLeader'],
+                        'canDeleteTickets' => in_array(1, $this->privileges['canDeleteTickets']),
+                        'ticket' => [
+                            'body' => $ticket['body'],
+                            'category_id' => $ticket['category_id']
+                        ],
+                        'category_name' => $category_name,
+                        'categories' => $categories,
+                        'lang' => $lang,
+                        'badges' => $badges
+                    ];
 
-                // load view
-                $this->loadView('ticket_edit', $data);
+                    // load view
+                    $this->loadView('ticket_edit', $data);
+                }
             }
         }
     }
@@ -293,55 +325,102 @@ class TicketsController extends Controller
 
             if ($data['canDeleteTickets'] && isset($_POST['delete_ticket'])) {
                 if ($this->ticketModel->deleteTicket($id)) {
+                    // log action
+                    $logAction = $_SESSION['user_name'] . ' deleted ticket (ID: ' . $id . ').';
+                    $logData = [
+                        'type' => 'Ticket',
+                        'action' => $logAction
+                    ];
+                    $this->logModel->adminLog($logData);
+
                     flashMessage('success', 'Ticket has been successfully deleted!');
                     redirect('/tickets');
                     unset($_POST);
                 } else {
                     die('Something went wrong');
                 }
-            }
-
-            if (isset($_POST['close_ticket'])) {
+            } else if ($data['canCloseTickets'] && isset($_POST['close_ticket'])) {
                 $status = 'Closed';
                 $closed_by = $_SESSION['user_id'];
 
                 if ($this->ticketModel->updateStatus($id, $status, $closed_by)) {
+                    // log action
+                    $logAction = $_SESSION['user_name'] . ' closed ticket (ID: ' . $id . ').';
+                    $logData = [
+                        'type' => 'Ticket',
+                        'action' => $logAction
+                    ];
+                    $this->logModel->adminLog($logData);
+
                     flashMessage('success', 'Ticket has been successfully closed!');
                     redirect('/tickets/view/' . $id);
                     unset($_POST);
                 } else {
                     die('Something went wrong');
                 }
-            }
-
-            if (isset($_POST['open_ticket'])) {
+            } else if ($data['canCloseTickets'] && isset($_POST['open_ticket'])) {
                 if ($this->ticketModel->updateStatus($id, 'Open')) {
-                    flashMessage('success', 'Ticket has been successfully closed!');
+                    // log action
+                    $logAction = $_SESSION['user_name'] . ' opened ticket (ID: ' . $id . ').';
+                    $logData = [
+                        'type' => 'Ticket',
+                        'action' => $logAction
+                    ];
+                    $this->logModel->adminLog($logData);
+
+                    flashMessage('success', 'Ticket has been successfully opened!');
                     redirect('/tickets/view/' . $id);
                     unset($_POST);
                 } else {
                     die('Something went wrong');
                 }
-            }
-
-            if ($data['canCloseTickets'] && isset($_POST['change_category'])) {
+            } else if ($data['canCloseTickets'] && isset($_POST['change_category'])) {
                 // filter post
                 $_POST = filter_input_array(INPUT_POST, FILTER_SANITIZE_STRING);
                 $cID = $_POST['new_category_id'];
                 // change category
                 if ($this->ticketModel->changeCategory($cID, $id)) {
+                    // log action
+                    $logAction = $_SESSION['user_name'] . ' changed ticket (ID: ' . $id . ') category to ' . $cID . '.';
+                    $logData = [
+                        'type' => 'Ticket',
+                        'action' => $logAction
+                    ];
+                    $this->logModel->adminLog($logData);
+
                     flashMessage('success', 'Ticket category has been successfully changed.');
                     redirect('/tickets/view/' . $id);
                 } else {
                     die('Something went wrong');
                 }
-            }
+            } else if ($data['canCloseTickets'] && isset($_POST['needs_owner'])) {
+                if ($this->ticketModel->updateStatus($id, 'Needs Owner Involvement')) {
+                    // log action
+                    $logAction = $_SESSION['user_name'] . ' changed ticket (ID: ' . $id . ') category to: Needs Owner Involvement.';
+                    $logData = [
+                        'type' => 'Ticket',
+                        'action' => $logAction
+                    ];
+                    $this->logModel->adminLog($logData);
 
-            if ($data['canDeleteTReplies'] && isset($_POST['delete_reply'])) {
+                    flashMessage('success', 'Ticket status has been successfully updated.');
+                    redirect('/tickets/view/' . $id);
+                } else {
+                    die('Something went wrong.');
+                }
+            } else if ($data['canDeleteTReplies'] && isset($_POST['delete_reply'])) {
                 // filter post
                 $_POST = filter_input_array(INPUT_POST, FILTER_SANITIZE_STRING);
                 $reply_id = $_POST['reply_id'];
                 if ($this->ticketModel->deleteReply($reply_id)) {
+                    // log action
+                    $logAction = $_SESSION['user_name'] . ' deleted reply #' . $reply_id . ' from ticket (ID: ' . $id . ').';
+                    $logData = [
+                        'type' => 'Ticket',
+                        'action' => $logAction
+                    ];
+                    $this->logModel->adminLog($logData);
+
                     flashMessage('success', 'Reply has been successfully deleted.');
                     redirect('/tickets/view/' . $id);
                 } else {
